// src/app/features/shared/machinery/models/machinery.model.ts
export interface Machinery {
    id: number;
    name: string;
    model: string;
    brand: string;
    serial_number: string;
    production_capacity: number;
    recommendations: string;
    status: number;
    user_creator: number;
    user_updater: number;
    created_at: string;
    updated_at: string;
  }